Refactoring: Extract app building logic into AppBuilder type
